pyproject.toml
[tool.poetry]
name = "repl_nix_lab05-stephanie-vicky"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
pil-compat = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
